requirements.txt
picsellia
torch
ultralytics
python-decouple